refactor(ffi): delete unneeded files, move clojure interaction to ffi
