--- src/jtests/mod.rs
@@ -1,1 +1,0 @@
-mod wr;
